PartA
