+++ GalgameManager/Services/ActivationService.cs
@@ -150,7 +150,6 @@ public class ActivationService : IActivationService
     private async Task StartupAsync() 
     {
         if (IsRestart() == false) App.SetWindowMode(WindowMode.Normal);
-        await _galgameCollectionService.StartAsync();
         if (IsRestart() == false) _pvnService.Startup();
         if (IsRestart() == false) await _updateService.UpdateSettingsBadgeAsync();
         await _appCenterService.StartAsync();
