+++ GalgameManager/Services/FilterService.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.ObjectModel;
+﻿using System.Collections;
+using System.Collections.ObjectModel;
 using GalgameManager.Contracts.Services;
 using GalgameManager.Enums;
 using GalgameManager.Helpers;
@@ -93,7 +94,7 @@ public class FilterService : IFilterService
             str = str[..(str.LastIndexOf('/') - 1)];
         await Task.Run((async Task() =>
         {
-            List<Galgame> games = (App.GetService<IGalgameCollectionService>() as GalgameCollectionService)!.Galgames;
+            IList<Galgame> games = (App.GetService<IGalgameCollectionService>() as GalgameCollectionService)!.Galgames;
             IEnumerable<CategoryGroup> categoryGroups = await App.GetService<ICategoryService>().GetCategoryGroupsAsync();
             //Category
             HashSet<string> addedCategories = new();
