+++ GalgameManager/Views/Control/GalgameSearchAutoSuggestBox.xaml
@@ -1,0 +1,16 @@
+<UserControl
+    x:Class="GalgameManager.Views.Control.GalgameSearchAutoSuggestBox"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <AutoSuggestBox x:Uid="HomePage_Search" AllowFocusOnInteraction="True" MinWidth="250"
+                    Text="{x:Bind SearchKey, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
+                    ItemsSource="{x:Bind SearchSuggestions, Mode=OneWay}" 
+                    TextChanged="AutoSuggestBox_OnTextChanged" 
+                    QuerySubmitted="AutoSuggestBox_OnQuerySubmitted">
+        
+    </AutoSuggestBox>
+</UserControl>
